requirements.txt
ansible==2.5.2
appdirs==1.4.3
asn1crypto==0.22.0
boto3=1.7.10
botocore==1.5.39
cffi==1.10.0
cryptography==1.8.1
docutils==0.13.1
idna==2.5
Jinja2==2.9.6
jmespath==0.9.2
MarkupSafe==1.0
packaging==16.8
paramiko==2.1.2
pyasn1==0.2.3
pycparser==2.17
pycrypto==2.6.1
pyparsing==2.2.0
python-dateutil==2.6.0
PyYAML==3.12
s3transfer==0.1.10
six==1.10.0
couchdb
requests
tweepy==3.5.0